Refactor the search, products, and orders pages.
