+++ components/orders/index.ts
@@ -1,0 +1,2 @@
+export * from './AddressInfo';
+export * from './OrderDetail';
