+++ pages/orders/history.tsx
@@ -3,11 +3,16 @@ import NextLink from 'next/link';
 import { getSession } from 'next-auth/react';
 
 import { Typography, Grid, Chip, Link } from '@mui/material';
-import { DataGrid, GridColDef, GridRenderCellParams } from '@mui/x-data-grid';
+import {
+  DataGrid,
+  GridColDef,
+  GridRenderCellParams,
+  GridToolbar,
+} from '@mui/x-data-grid';
 
-import { ShopLayout } from '../../components/layouts';
 import { dbOrders } from '@/database';
 import { IOrder } from '@/interfaces';
+import { ShopLayout } from '../../components/layouts';
 
 const columns: GridColDef[] = [
   { field: 'id', headerName: 'ID', width: 100 },
@@ -33,7 +38,11 @@ const columns: GridColDef[] = [
     sortable: false,
     renderCell: (params: GridRenderCellParams) => {
       return (
-        <NextLink href={`/orders/${params.row.orderId}`} passHref legacyBehavior>
+        <NextLink
+          href={`/orders/${params.row.orderId}`}
+          passHref
+          legacyBehavior
+        >
           <Link underline="always">Ver orden</Link>
         </NextLink>
       );
@@ -42,7 +51,7 @@ const columns: GridColDef[] = [
 ];
 type Props = {
   orders: IOrder[];
-}
+};
 
 const HistoryPage: NextPage<Props> = ({ orders }) => {
   const rows = orders.map((order, index) => {
@@ -63,9 +72,22 @@ const HistoryPage: NextPage<Props> = ({ orders }) => {
         Historial de ordenes
       </Typography>
 
-      <Grid container className='fadeIn'>
-        <Grid item xs={12} sx={{ height: 440, width: '100%' }}>
-          <DataGrid rows={rows} columns={columns} autoPageSize />
+      <Grid container className="fadeIn">
+        <Grid item xs={12} sx={{ height: 460, width: '100%' }}>
+          <DataGrid
+            rows={rows}
+            columns={columns}
+            localeText={{
+              toolbarDensity: 'Size',
+              toolbarDensityLabel: 'Size',
+              toolbarDensityCompact: 'Small',
+              toolbarDensityStandard: 'Medium',
+              toolbarDensityComfortable: 'Large',
+            }}
+            slots={{
+              toolbar: GridToolbar,
+            }}
+          />
         </Grid>
       </Grid>
     </ShopLayout>
